Segfault caused by multiple versions of yui installed
Installing yast2-ycp-ui-bindings-devel brought libyui8 with it, and
having both libyui7 and 8 installed at the same time causes libtool to link
to both libs, which segfaults on exit. Not having yast2-ycp-ui-bindings-devel
installed means we don't get the symlink we need to build against
libpy2UI.so, and we also don't get the YCP_UI.h header file. We'll
have to take care of this in packaging (or get the yast folks to fix this).